backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis and Celery
redis==5.0.1
celery==5.3.4

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# AI/ML
langchain==0.1.0
langchain-community==0.0.10
anthropic==0.7.8
openai==1.3.7
faiss-cpu==1.9.0.post1
spacy==3.7.2
sentence-transformers==2.2.2

# HTTP and API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# CORS
python-cors==1.0.0

# WebSocket
websockets==12.0

# Monitoring
prometheus-client==0.19.0

# Logging
loguru==0.7.2

# Environment
python-dotenv==1.0.0
